+++ packages/liveblocks-react-ui/src/components/Comment.tsx
@@ -222,11 +222,11 @@ export function CommentMention({
     case "user":
       return (
         <CommentPrimitive.Mention
-          className={cn("lb-comment-mention", className)}
+          className={cn("lb-mention lb-comment-mention", className)}
           data-self={mention.id === currentId ? "" : undefined}
           {...props}
         >
-          {MENTION_CHARACTER}
+          <span className="lb-mention-symbol">{MENTION_CHARACTER}</span>
           <User userId={mention.id} />
         </CommentPrimitive.Mention>
       );